requirements.txt
flask==2.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
